refactor/states: reduce the number of state traits
